requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
aio-pika==9.5.7
anyio==4.12.1
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
beanie==2.0.1
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.128.2
fastapi-pagination==0.15.9
h11==0.16.0
idna==3.11
lazy-model==0.4.0
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.22
redis==7.2.1
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
pandas==2.2.3
openpyxl==3.1.5
slowapi==0.1.9
tenacity==9.1.4
